+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 0.12.0-dev
 
+## 0.11.0
+
 ### Packaging
 
 - Minimum Rust version has been bumped to 1.57.0